tournament-service/requirements.txt
django-filter>=23.5
django-extensions>=3.2.3
django-polymorphic>=3.1.0
django-treebeard>=4.7.1
django-mptt>=0.15.0  # For hierarchical tournament structures
pandas>=2.1.3  # For data analysis and standings calculations
django-guardian>=2.4.0  # For object-level permissions
celery>=5.3.6  # For asynchronous tasks
redis>=5.0.1  # For Celery broker
drf-yasg>=1.21.7  # Swagger/OpenAPI documentation 